+++ src/learning/custom/custom_ltr.cc
@@ -53,7 +53,7 @@ void CustomLTR::learn(
     std::shared_ptr<quickrank::data::Dataset> training_dataset,
     std::shared_ptr<quickrank::data::Dataset> validation_dataset,
     std::shared_ptr<quickrank::metric::ir::Metric> scorer,
-    unsigned int partial_save, const std::string output_basename) {
+    size_t partial_save, const std::string output_basename) {
 
   // Do some initialization
   preprocess_dataset(training_dataset);
@@ -67,7 +67,7 @@ void CustomLTR::learn(
   Score* validation_scores = new Score[validation_dataset->num_instances()];
 
   // set scores equal to fixed value
-  for (unsigned int i = 0; i < training_dataset->num_instances(); i++)
+  for (size_t i = 0; i < training_dataset->num_instances(); i++)
     training_scores[i] = FIXED_SCORE;
 
   MetricScore metric_on_training = scorer->evaluate_dataset(training_dataset,
@@ -75,7 +75,7 @@ void CustomLTR::learn(
 
   std::cout << *scorer << " on training: " << metric_on_training << std::endl;
 
-  for (unsigned int i = 0; i < validation_dataset->num_instances(); i++)
+  for (size_t i = 0; i < validation_dataset->num_instances(); i++)
     validation_scores[i] = FIXED_SCORE;
 
   MetricScore metric_on_validation = scorer->evaluate_dataset(
@@ -92,7 +92,7 @@ void CustomLTR::learn(
 
 // assumes vertical dataset
 Score CustomLTR::score_document(const quickrank::Feature* d,
-                                const unsigned int offset) const {
+                                const size_t offset) const {
   return FIXED_SCORE;
 }
